+++ src/main/java/net/libz/network/LibzServerPacket.java
@@ -6,12 +6,8 @@ import net.libz.network.packet.ConfigPacket;
 import net.libz.network.packet.MousePacket;
 import net.libz.util.ConfigHelper;
 import net.minecraft.server.network.ServerPlayerEntity;
-import net.minecraft.util.Identifier;
 
 public class LibzServerPacket {
-
-    public static final Identifier SET_MOUSE_POSITION = Identifier.of("libz", "set_mouse_position");
-    public static final Identifier SYNC_CONFIG_PACKET = Identifier.of("libz", "sync_config");
 
     public static void init() {
         PayloadTypeRegistry.playS2C().register(ConfigPacket.PACKET_ID, ConfigPacket.PACKET_CODEC);